Update README.md
Added a static badge. Let me know if you prefer to have any other static badges. I will add dynamic badges as soon as we launch into public beta.

The examples can be added only after we confirm the issue we found during yesterday's installation.

I was also thinking of adding a contributors table in the readme only (See Yuso's readme for example).
Also, consider adding `How to contribute` in the readme instead of a separate file to maximize the utilisation of the initial footfall.